move deprecated type TempTableDdlTransactionHandling close to its clients
